requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
channels==4.1.0
channels-redis==4.2.0
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
msgpack==1.1.0
packaging==24.1
psycopg2==2.9.10
PyJWT==2.8.0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
sqlparse==0.5.0
typing-extensions==4.12.2
uritemplate==4.1.1
